O(nlog(n)) time O(1) space
Given a number i, if the number of elements in nums no more than i is smaller than or equal to i, then the duplicated number should be larger than i. Otherwise, the duplicated number should be smaller than or equal to i. Therefore, we can use binary search to find the typical i which refers to the duplicated number.

Be careful about how to set left and right and what to return.