O(n^2) time, O(1) space, 80ms
This version should save some time than the first version although the runtime is still 80ms on passing 21 test cases. This is because when the current node is already bigger than all the previous ones, we do not need to do the insertion operation. 

Note that if we did not allocate a new node for dummy (write ListNode* dummy; instead of ListNode* dummy=new ListNode(INT_MIN);), we would not pass the test cases.

Also note that if we do not use the node pointer current but directly use pre->next, the runtime would be 94ms.

Also note that if we write num=current->val and use num instead of current->val in the while condition, the runtime becomes 92ms.