O(n) time, O(1) space, 580ms
If newInterval is smaller than the current interval, insert newInterval and then the current interval into the result.
If newInterval is overlapped with the current interval, update the newInterval by merging them together.
If newInterval is bigger than the current interval, insert the current interval only.
We use a flag to record whether the newInterval has already been inserted or not. For the above three conditions, only in the first condition, the flag is set to be true. If the flag is false after going through intervals, we insert the newInterval into the result.